move non-partition modules to utils folder
move partition/json.py back
