+++ unstructured/nlp/partition.py
@@ -1,6 +1,6 @@
 # flake8: noqa
 from unstructured.partition.pdf import partition_pdf  # noqa
-from unstructured.partition.text_type import (  # noqa
+from unstructured.partition.utils.text_type import (  # noqa
     is_bulleted_text,
     is_possible_narrative_text,
     is_possible_title,
