+++ unstructured/partition/doc.py
@@ -7,13 +7,13 @@ from typing import IO, Any, Optional
 from unstructured.chunking import add_chunking_strategy
 from unstructured.documents.elements import Element, process_metadata
 from unstructured.file_utils.filetype import FileType, add_metadata_with_filetype
-from unstructured.partition.common import (
+from unstructured.partition.docx import partition_docx
+from unstructured.partition.utils.common import (
     convert_office_doc,
     exactly_one,
     get_last_modified_date,
     get_last_modified_date_from_file,
 )
-from unstructured.partition.docx import partition_docx
 
 
 @process_metadata()
