+++ unstructured/partition/image.py
@@ -5,12 +5,12 @@ from typing import Any, Optional
 from unstructured.chunking import add_chunking_strategy
 from unstructured.documents.elements import Element, process_metadata
 from unstructured.file_utils.filetype import add_metadata
-from unstructured.partition.common import exactly_one
-from unstructured.partition.lang import (
+from unstructured.partition.pdf import partition_pdf_or_image
+from unstructured.partition.utils.common import exactly_one
+from unstructured.partition.utils.constants import PartitionStrategy
+from unstructured.partition.utils.lang import (
     check_language_args,
 )
-from unstructured.partition.pdf import partition_pdf_or_image
-from unstructured.partition.utils.constants import PartitionStrategy
 
 
 @process_metadata()
