+++ unstructured/partition/pptx.py
@@ -35,18 +35,18 @@ from unstructured.documents.elements import (
     process_metadata,
 )
 from unstructured.file_utils.filetype import FileType, add_metadata_with_filetype
-from unstructured.partition.common import (
+from unstructured.partition.utils.common import (
     convert_ms_office_table_to_text,
     get_last_modified_date,
     get_last_modified_date_from_file,
 )
-from unstructured.partition.lang import apply_lang_metadata
-from unstructured.partition.text_type import (
+from unstructured.partition.utils.constants import PartitionStrategy
+from unstructured.partition.utils.lang import apply_lang_metadata
+from unstructured.partition.utils.text_type import (
     is_email_address,
     is_possible_narrative_text,
     is_possible_title,
 )
-from unstructured.partition.utils.constants import PartitionStrategy
 from unstructured.utils import is_temp_file_path, lazyproperty
 
 DETECTION_ORIGIN = "pptx"
