+++ unstructured/partition/text.py
@@ -28,13 +28,13 @@ from unstructured.file_utils.encoding import read_txt_file
 from unstructured.file_utils.filetype import FileType, add_metadata_with_filetype
 from unstructured.nlp.patterns import PARAGRAPH_PATTERN, UNICODE_BULLETS_RE
 from unstructured.nlp.tokenize import sent_tokenize
-from unstructured.partition.common import (
+from unstructured.partition.utils.common import (
     exactly_one,
     get_last_modified_date,
     get_last_modified_date_from_file,
 )
-from unstructured.partition.lang import apply_lang_metadata
-from unstructured.partition.text_type import (
+from unstructured.partition.utils.lang import apply_lang_metadata
+from unstructured.partition.utils.text_type import (
     is_bulleted_text,
     is_email_address,
     is_possible_narrative_text,
