+++ unstructured/partition/xlsx.py
@@ -25,9 +25,12 @@ from unstructured.documents.elements import (
     process_metadata,
 )
 from unstructured.file_utils.filetype import FileType, add_metadata_with_filetype
-from unstructured.partition.common import get_last_modified_date, get_last_modified_date_from_file
-from unstructured.partition.lang import apply_lang_metadata
-from unstructured.partition.text_type import (
+from unstructured.partition.utils.common import (
+    get_last_modified_date,
+    get_last_modified_date_from_file,
+)
+from unstructured.partition.utils.lang import apply_lang_metadata
+from unstructured.partition.utils.text_type import (
     is_bulleted_text,
     is_possible_narrative_text,
     is_possible_numbered_list,
